feat: Add helper module
